A long time ago, and yet somehow in the future...
                            Chapter IV

It was a dark time in the Subversion tree. The Federation, headed by
the dark Sith Lord Vader, managed to break into the Subversion tree of
the small band of rebels and submit a patch to build_gcc under the
pseudonym "void".

All was not lost! Jedi Devang noticed this breach and informed our
hero that this was in error! And that the LLVM core was to be compiled
separately from the LLVM-GCC front-ends.

The patch will hence be reverted...plus a couple of new things to help
it along...

[Cue star field and space ship]

llvm-svn: 43759