Fix fallout from recent commits. This fixes PR1785
Folks, when syncing to Apple gcc, please, do remember, that some files
are nullified in the gcc/config directory (mostly for targets Apple doesn't care), 
don't propagate this to llvm-gcc. This stuff is hard to catch, because it doesn't
lead to any errors during gcc compilation, only resulting binaries are broken.

llvm-svn: 43937