backporting r46308:
The initialization for _OBJC_IVAR_$_<classname>.b<integer> symbols are
dropped
for -O0. As it turns out, the LLVM variable is created at -O0 before the
initialization is set to the correct value. It defaults to 0, so these
use the
zero filled directive, which is wrong.

Reset the initialization on LLVM variables when needed.

objc2-bitfield-abi-1.m is a testcase that exercises this.

llvm-svn: 46311