Fixed bug where BU used the callgraph to discover callees for inlining. This was an artifact from the work on partial inlining (see r98921), and enabled situations where BU would try to inline the graph for an unvisited function. (Which can easily result in all kinds of badness)
This fixes the infinite recursion in calculateGraph triggered by sqlite3.

Finally, I believe this means we can also remove the eraseCS stuff, since if we
get callees for a site it is because we are able to resolve it (and so should
remove it).  Will visit this later, I believe present behavior isn't *wrong*.

llvm-svn: 112150