Move the code that propogates information to the globals graph to be after the Incomplete/External flag calculation.
This ensures that the globals graph does not get I, in the cases
when the graph itself does not mark I on the same node in its
graph. 

This was happening when a global was merged with a formal 
parameter to the function, which was I till the end of TD
for that graph.

llvm-svn: 118664