As MCJIT does not implement properly runFunction, generate a function gate to interact with JITted code from C++ code for each llvm signature.
Basically, each J3MethodType as an associated J3LLVMSignature. The
J3LLVMSignature contains a llvm function type and a gate. When a
function f with a given signature S is JITted, verifies whether the
gate for S was already generated. If it is not the case, generate the
gate for S when we generate f. As now, a Java object is represented by
a J3Object*, whatever the type of the Java object, the number of gates
should not be too crazy.

The gate takes in charge the transformation between a J3ObjectHandle
and a J3Object and totally hides J3Object to the caller.

llvm-svn: 198282