Remove dead projects: core, sample, support, llvm-top, website.
core and support:

Was added in in 2007 (r39913/r39914) to start to split off some of
llvm into smaller libraries. Was abandoned by December 2007, last
commit to support is r44863, core never got a commit.

llvm-top:

Was added in 2007 in r37869, as a convenience to check out and build
multiple llvm projects, as part of the above plan to split things
apart. Last touched in 2009, r62840. As far as I know, completely
non-functional today.

sample:

Was added in 2007 in r39940, and last touched late 2007 at r40403. Not
actually a useful sample.

website:

Was added in 2007, r37886. Last update in 2007, r40389.
Note: the real website lives in "www", not here.

llvm-svn: 286375